// Client_dapp_bd4opem/frontEnd/src/app/models/contract.ts
/**
 * BD4OPEM Account Management module API Specification
 * This document contains the formal specification of the interfaces with Account and Contracting management in the BD4OPEM Marketplace.
 *
 * OpenAPI spec version: 0.0.25
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import internal from 'stream';
import { ContractBusinessParameters } from './contractBusinessParameters';

/**
 * The contract claims to be issued in the Contract VC
 */
export interface Contract { 
    /**
     * The Contract UUID identity
     */
    contractID?: string;
    businessParmaeters?: ContractBusinessParameters;
    version?: string;
    /**
     * The status of the contract.
     */
    status?: string;
    /**
     * Contract claims specific to a Service User.
     */
    serviceParameters?: any;
    dataParameters?: Array<{
        oganizationId	: string;
        accrualPeriod: String;
        resourceEndpoint: string;
        offeringType: string;
        granularity: string;

        meteringPoints: Array<any>;
        meteringPointNo: number;
        attributes: Array<any>;
        type: string;
        data_source_id: string;
    }>;
}